fix: properly default cjs to .cjs as extension
BREAKING CHANGE: Though previously already intended for cjs to get `.cjs`
as file extension, that was never actually properly applied. Because the
CommonJS extension will from this version forward actually properly be
`.cjs` I am releasing this a breaking change in case anyone
expected it to be `.js`. If you want to upgrade and retain `.js` pass
`{ cjsExtension: 'js' }` as options to this plugin.